+++ src/services/config_extraction.rs
@@ -256,6 +256,10 @@ impl ConfigExtractionService for ConfigExtractionServiceImpl {
                 .evg_config_utils
                 .lookup_build_variant_expansion(UNIQUE_GEN_SUFFIX_EXPANSION, variant);
         }
+        let num_tasks = self
+            .evg_config_utils
+            .get_gen_task_var(task_def, "num_tasks")
+            .map(|str| str.parse().unwrap());
 
         Ok(ResmokeGenParams {
             task_name,
@@ -297,6 +301,7 @@ impl ConfigExtractionService for ConfigExtractionServiceImpl {
             pass_through_vars: self.evg_config_utils.get_gen_task_vars(task_def),
             platform,
             gen_task_suffix,
+            num_tasks,
         })
     }